split out rules logic and add a grabber level for the rules
